feat: Implement query plans for sqlserver dialect

+++ queryeer-backend/backend-contract/src/main/java/com/queryeer/backend/contract/graph/GraphEdgeStyle.java
@@ -1,5 +1,5 @@
 package com.queryeer.backend.contract.graph;
 
-public record GraphEdgeStyle(String shape, String color, Integer width, Boolean dash, String markerEnd)
+public record GraphEdgeStyle(String shape, String color, Double width, Boolean dash, String markerEnd)
 {
 }
